+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * Fix `Shorthand` to not report shorthanded values with `!important` as having
   a shorthand length not specified in `allowed_shorthands`
 * Fix `UnnecessaryMantissa` to ignore decimal values in URL literals
+* Fix `PlaceholderInExtend` to report comma sequences starting with a
+  placeholder
 
 ## 0.49.0